15.03.26 Если API вернуло None

+++ services_weather/weather_coords/coordinates.py
@@ -15,23 +15,33 @@ async def handle_user_location(message: Message):
     """
     logger = logging.getLogger(__name__)
 
+    # 1. Отправляет сообщение о начале процесса.
     wait_msg = await message.answer("📡 Связываюсь со спутником...")
+
     try:
         # Получает широту и долготу.
         lat = message.location.latitude
         lon = message.location.longitude
         logger.info(f"lat и lon получены. {lat}, {lon}")
 
-        # Вызывает погоду - JSON по координатам.
+        # Вызывает погоду через API - JSON по координатам.
         weather_dict = get_weather_by_coords(lat, lon)
+
+        # ПРОВЕРКА: Если API вернуло None (ошибка в get_weather_by_coords).
+        if not weather_dict:
+            raise Exception("API вернуло пустой ответ")
 
         # Превращает JSON в красивый текст.
         await message.answer(weather_dict, reply_markup=main_keyboard)
         await wait_msg.delete()
 
-
     except Exception as e:
         logger.exception(f"Ошибка в обработчике локации: {e}")
+        # Пытаемся удалить сообщение об ожидании, если оно еще есть.
+        try:
+            await wait_msg.delete()
+        except: # noqa
+            pass
 
         # Если ошибка, предлагаем попробовать еще раз через инлайн-кнопку.
         await message.answer(
